Updated Filters to allow std::optional.
Previously, a filter's Get and Set methods would only be called
after derefencing a std::optional value. This style filter is
still supported.

If the filter's Get/Set method accepts std::optional, std::optional
values will be passed as-is.

Added ChoiceMuxer to allow choices in a Select object to be changed
by a mux value.